refactoring folling pep8
